requirements.txt
flask
scikit-learn
numpy
requests
beautifulsoup4
reportlab
waitress
duckduckgo-search
google-genai